feat: add shared WebSocket types and consolidate API contracts
- Move ClientMessage, ServerMessage, PresenceStatus from server to shared crate
- Add error_codes module with WS protocol constants
- Use Vec<u8> with base64 serde for encrypted WS message fields
- Add dm_channel_id to MessageResponse with skip_serializing_if
- Server ws/types.rs now re-exports from shared

Plan: section-11-shared-types.md
